pyproject.toml
[tool.poetry]
name = "lotto"
version = "1.0.0"
description = "url단축 서비스 api입니다."
authors = ["ShortUrl"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2"
pytest = "8.1.1"
pytest-django = "4.8.0"
djangorestframework = "3.14.0"
psycopg2-binary = "2.9.9"
poetry = "1.8.2"
python-dotenv = "1.0.1"
loguru = "0.7.2"
drf-yasg = "1.21.7"
pre-commit = "3.6.2"
isort = "5.13.2"
black = "24.3.0"






[tool.pytest.ini_options]
addopts = "-ra -q"
python_files = ["test_*.py"]
markers = [
    "unit"
]

